requirements.txt
streamlit
google-generativeai
vapi_server_sdk
requests
PyPDF2